correction cursor (cast) rajout support souris pour editfield back au lieu de cancel dans les menus reseau
git-svn-id: svn+ssh://src.fovea.cc/svn/flobopuyo/trunk@763 053fdb51-1f17-0410-be66-d00040120cd6

+++ gametools/GameCursor.cpp
@@ -34,7 +34,7 @@ void GameCursor::draw(SDL_Surface *screen)
     //nvy = (nvy * 0.5) * cursorSurface->h;
     float nvx = (-sin(anglerad) * 0.5) * cursorSurface->w;
     float nvy = (-cos(anglerad) * 0.5) * cursorSurface->h;
-    IIM_BlitRotatedSurfaceCentered(cursorSurface, angle, screen, blitX - nvx, blitY - nvy);
+    IIM_BlitRotatedSurfaceCentered(cursorSurface, (int)angle, screen, (int)(blitX - nvx), (int)(blitY - nvy));
 }
 
 void GameCursor::onEvent(GameControlEvent *event)
